bottom navbar, file management
